Support embedded etcd (#14333)

+++ internal/indexcoord/index_coord.go
@@ -50,6 +50,7 @@ import (
 	"github.com/milvus-io/milvus/internal/util/trace"
 	"github.com/milvus-io/milvus/internal/util/tsoutil"
 	"github.com/milvus-io/milvus/internal/util/typeutil"
+	clientv3 "go.etcd.io/etcd/client/v3"
 )
 
 // make sure IndexCoord implements types.IndexCoord
@@ -79,7 +80,8 @@ type IndexCoord struct {
 
 	idAllocator *allocator.GlobalIDAllocator
 
-	kv kv.BaseKV
+	etcdCli *clientv3.Client
+	kv      kv.BaseKV
 
 	metaTable   *metaTable
 	nodeManager *NodeManager
@@ -129,17 +131,19 @@ func (i *IndexCoord) Register() error {
 			log.Fatal("failed to stop server", zap.Error(err))
 		}
 		// manually send signal to starter goroutine
-		syscall.Kill(syscall.Getpid(), syscall.SIGINT)
+		if i.session.TriggerKill {
+			syscall.Kill(syscall.Getpid(), syscall.SIGINT)
+		}
 	})
 	return nil
 }
 
 func (i *IndexCoord) initSession() error {
-	i.session = sessionutil.NewSession(i.loopCtx, Params.IndexCoordCfg.MetaRootPath, Params.IndexCoordCfg.EtcdEndpoints)
+	i.session = sessionutil.NewSession(i.loopCtx, Params.IndexCoordCfg.MetaRootPath, i.etcdCli)
 	if i.session == nil {
 		return errors.New("failed to initialize session")
 	}
-	i.session.Init(typeutil.IndexCoordRole, Params.IndexCoordCfg.Address, true)
+	i.session.Init(typeutil.IndexCoordRole, Params.IndexCoordCfg.Address, true, true)
 	Params.BaseParams.SetLogger(i.session.ServerID)
 	return nil
 }
@@ -160,10 +164,7 @@ func (i *IndexCoord) Init() error {
 		}
 
 		connectEtcdFn := func() error {
-			etcdKV, err := etcdkv.NewEtcdKV(Params.IndexCoordCfg.EtcdEndpoints, Params.IndexCoordCfg.MetaRootPath)
-			if err != nil {
-				return err
-			}
+			etcdKV := etcdkv.NewEtcdKV(i.etcdCli, Params.IndexCoordCfg.MetaRootPath)
 			metakv, err := NewMetaTable(etcdKV)
 			if err != nil {
 				return err
@@ -207,12 +208,7 @@ func (i *IndexCoord) Init() error {
 
 		//init idAllocator
 		kvRootPath := Params.IndexCoordCfg.KvRootPath
-		etcdKV, err := tsoutil.NewTSOKVBase(Params.IndexCoordCfg.EtcdEndpoints, kvRootPath, "index_gid")
-		if err != nil {
-			log.Error("IndexCoord TSOKVBase initialize failed", zap.Error(err))
-			initErr = err
-			return
-		}
+		etcdKV := tsoutil.NewTSOKVBase(i.etcdCli, kvRootPath, "index_gid")
 
 		i.idAllocator = allocator.NewGlobalIDAllocator("idTimestamp", etcdKV)
 		if err := i.idAllocator.Initialize(); err != nil {
@@ -304,6 +300,10 @@ func (i *IndexCoord) Stop() error {
 	i.session.Revoke(time.Second)
 
 	return nil
+}
+
+func (i *IndexCoord) SetEtcdClient(etcdClient *clientv3.Client) {
+	i.etcdCli = etcdClient
 }
 
 // UpdateStateCode updates the component state of IndexCoord.
