+++ internal/proxy/proxy.go
@@ -38,6 +38,7 @@ import (
 	"github.com/milvus-io/milvus/internal/util/paramtable"
 	"github.com/milvus-io/milvus/internal/util/sessionutil"
 	"github.com/milvus-io/milvus/internal/util/typeutil"
+	clientv3 "go.etcd.io/etcd/client/v3"
 	"go.uber.org/zap"
 )
 
@@ -67,6 +68,7 @@ type Proxy struct {
 
 	stateCode atomic.Value
 
+	etcdCli    *clientv3.Client
 	rootCoord  types.RootCoord
 	indexCoord types.IndexCoord
 	dataCoord  types.DataCoord
@@ -116,7 +118,9 @@ func (node *Proxy) Register() error {
 		if err := node.Stop(); err != nil {
 			log.Fatal("failed to stop server", zap.Error(err))
 		}
-		syscall.Kill(syscall.Getpid(), syscall.SIGINT)
+		if node.session.TriggerKill {
+			syscall.Kill(syscall.Getpid(), syscall.SIGINT)
+		}
 	})
 	// TODO Reset the logger
 	//Params.initLogCfg()
@@ -125,11 +129,11 @@ func (node *Proxy) Register() error {
 
 // initSession initialize the session of Proxy.
 func (node *Proxy) initSession() error {
-	node.session = sessionutil.NewSession(node.ctx, Params.ProxyCfg.MetaRootPath, Params.ProxyCfg.EtcdEndpoints)
+	node.session = sessionutil.NewSession(node.ctx, Params.ProxyCfg.MetaRootPath, node.etcdCli)
 	if node.session == nil {
 		return errors.New("new session failed, maybe etcd cannot be connected")
 	}
-	node.session.Init(typeutil.ProxyRole, Params.ProxyCfg.NetworkAddress, false)
+	node.session.Init(typeutil.ProxyRole, Params.ProxyCfg.NetworkAddress, false, true)
 	Params.ProxyCfg.ProxyID = node.session.ServerID
 	Params.BaseParams.SetLogger(Params.ProxyCfg.ProxyID)
 	return nil
@@ -423,6 +427,11 @@ func (node *Proxy) lastTick() Timestamp {
 // AddCloseCallback adds a callback in the Close phase.
 func (node *Proxy) AddCloseCallback(callbacks ...func()) {
 	node.closeCallbacks = append(node.closeCallbacks, callbacks...)
+}
+
+// SetEtcdClient sets etcd client for proxy.
+func (node *Proxy) SetEtcdClient(client *clientv3.Client) {
+	node.etcdCli = client
 }
 
 // SetRootCoordClient sets RootCoord client for proxy.
